requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-allauth==0.39.1
django-ckeditor==5.8.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-import-export==1.2.0
django-jet==1.0.8
django-js-asset==1.2.2
et-xmlfile==1.0.1
gunicorn==20.0.0
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.7
xlrd==1.2.0
xlwt==1.3.0
